requirements.txt
krakenex>=2.1.0
pykrakenapi>=0.3.1
requests>=2.31.0
python-dotenv>=1.0.0
